Cleaned up Robot and RoverService, fixed service shutdown.
The "click to stop" notification didn't work because:
-the intent it used actually starts the service and passes an action "stop"
 which was ignored
-once the intent action is actually handled, the service onDestroy failed to
 unregister the ioio and stop the AI.
-once the service onDestroy was implemented correctly, closing the activity
 and stopping the service separately have strange effects:
 -closing only the service makes the whole app invisble to running apps, while
  AI, gyros, and vision continue to run.
 -closing only the app causes it to be restarted.
so in the interest of sanity, it has been changed to a more convental "raise
associated activity" behavior, with a proper shutdown button on said activity.

The libs/android-support-v4.jar is for the compatible, non-deprecated
notification api. The new notifiation code is exactly the same except for the
changes noted above.